requirements.txt
annotated-types
anyio
bcrypt
certifi
click
dnspython
email_validator
fastapi
fastapi-cli
fastapi-cloud-cli
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
psycopg
psycopg-binary
psycopg-pool
psycopg2-binary
pydantic
pydantic_core
Pygments
PyJWT
python-dotenv
python-multipart
PyYAML
rich
rich-toolkit
rignore
sentry-sdk
shellingham
sniffio
starlette
typer
typing-inspection
typing_extensions
urllib3
uvicorn
watchfiles
websockets
aiohttp